feat: add readable time formatting and display transformation in ivy-fd
This commit introduces two new functions to the ivy-fd package:
- `ivy-fd-format-time-readable`: This function calculates the time
difference from a given time to the current time and formats it in a
human-readable string, such as "3 hours ago" or "2 days ago".
- `ivy-fd--read-file-display-transformer`: This function transforms the
display of filenames by appending the modification time in a readable
format and adjusting the display properties based on the file's type and
accessibility.

Additionally, the `ivy-fd-async` configuration is updated to use the new
display transformer function.